requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.5
blinker==1.6.3
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
cytoolz==0.12.1
distlib==0.3.6
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.1.1
fastapi==0.103.1
filelock==3.12.0
firebase-admin==6.2.0
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.3.3
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
hexbytes==0.3.1
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.2.0
MarkupSafe==2.1.3
msgpack==1.0.5
multidict==6.0.4
parsimonious==0.9.0
Pillow==10.0.1
pipenv==2023.6.2
platformdirs==3.5.1
proto-plus==1.22.3
protobuf==4.24.3
py-solc-x==1.1.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pyparsing==3.1.1
pyrsistent==0.19.3
pywin32==306
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
rsa==4.9
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
solc==0.0.0a0
starlette==0.27.0
toolz==0.12.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
virtualenv==20.23.0
virtualenv-clone==0.5.7
web3==6.5.0
websockets==11.0.3
Werkzeug==3.0.1
yarl==1.9.2
